move Manifest schemas outside of config schemas
